PT-145/PT-148 Persist in pipeline, improve database connection

+++ db-layer/src/test/java/com/novoda/github/reports/data/db/DbEventDataLayerTest.java
@@ -22,10 +22,10 @@ import static org.mockito.Mockito.when;
 
 public class DbEventDataLayerTest {
 
-    private static final int ANY_EVENT_ID = 1337;
-    private static final Integer ANY_EVENT_REPO_ID = 1;
-    private static final Integer ANY_EVENT_AUTHOR_ID = 2;
-    private static final Integer ANY_EVENT_OWNER_ID = 3;
+    private static final Long ANY_EVENT_ID = 1337L;
+    private static final Long ANY_EVENT_REPO_ID = 1L;
+    private static final Long ANY_EVENT_AUTHOR_ID = 2L;
+    private static final Long ANY_EVENT_OWNER_ID = 3L;
     private static final EventType ANY_EVENT_TYPE = EventType.ISSUE_COMMENT_ADD;
     private static final Date ANY_EVENT_DATE = new GregorianCalendar(2016, 4, 26, 18, 20).getTime();
     private DbEventDataLayer dataLayer;
